Refactor: changing customer data validation location

+++ src/main/java/com/electricitybill/service/impl/ClienteServiceImpl.java
@@ -3,10 +3,12 @@ package com.electricitybill.service.impl;
 import com.electricitybill.entity.ClienteEntity;
 import com.electricitybill.repository.ClienteRepository;
 import com.electricitybill.service.ClienteService;
+import jakarta.persistence.NoResultException;
 import lombok.AllArgsConstructor;
 
 import java.sql.SQLException;
 import java.util.List;
+import java.util.Objects;
 
 @AllArgsConstructor
 public class ClienteServiceImpl implements ClienteService {
@@ -15,11 +17,17 @@ public class ClienteServiceImpl implements ClienteService {
 
     @Override
     public List<ClienteEntity> findAll() {
+        if (clienteRepository.findAll().isEmpty()) {
+            throw new NoResultException("Search not found!");
+        }
         return clienteRepository.findAll();
     }
 
     @Override
     public ClienteEntity findById(int id) {
+        if (Objects.isNull(clienteRepository.findById(id))) {
+            throw new NoResultException("Individual search not found!");
+        }
         return clienteRepository.findById(id);
     }
 
@@ -30,11 +38,17 @@ public class ClienteServiceImpl implements ClienteService {
 
     @Override
     public boolean update(int id, ClienteEntity entity) throws SQLException {
+        if (Objects.isNull(clienteRepository.findById(id))) {
+            throw new NoResultException("Individual search not found!");
+        }
         return clienteRepository.update(id, entity);
     }
 
     @Override
     public boolean delete(int id) throws SQLException {
+        if (Objects.isNull(clienteRepository.findById(id))) {
+            throw new NoResultException("Individual search not found!");
+        }
         return clienteRepository.delete(id);
     }
 }
